fix #713341883 完善文档，调整days类型

+++ src/struct/MedicalInsuranceItem.php
@@ -80,7 +80,7 @@ class MedicalInsuranceItem extends Base
     /**
      * 天数
      *
-     * @var integer|null
+     * @var float|null
      */
-    public ?int $days = null;
+    public ?float $days = null;
 }
